Added the category to the list of expenses
fixes to the select list for categories in the edit/add expense modal:
- placeholder text displays
- if value is selected then it displays instead
- can save the value

Changed the expense to hold the actual category element instead of just an id for a category to make it easier to display the category name. It still saves to the json file as just an id.